freeswitch: use the installation path in the init script
The Freeswitch code from the original repository assumes
that it was installed via apt which means that the binary
is in /usr/bin and the configuration is in /etc.

This change uses Plivo's Freeswitch directory
instead: /usr/local/freeswitch

WARNING! I set WORKDIR for start-stop-daemon to $FS_INSTALLED_PATH,
because I don't know what it's supposed to be. The one that ships
with Ubuntu assumes it's in ~freeswitch(/var/lib/freeswitch).